Updated the data structures to store money as an integer in cents rather than a double as dollars to preserve accuracy, and will just be presented as a dollar amount at the UI level.

+++ Source/AllowanceApp.Api/DTO/AccountDTO.cs
@@ -3,7 +3,7 @@ using AllowanceApp.Api.Utilities;
 
 namespace AllowanceApp.Api.DTO
 {
-    public record AccountDTO(string Name, int ID, double Balance, List<AllowancePointDTO> Allowances, List<TransactionDTO> Transactions)
+    public record AccountDTO(string Name, int ID, int Balance, List<AllowancePointDTO> Allowances, List<TransactionDTO> Transactions)
     {
         public AccountDTO(Account account) : this
         (
