+++ Source/AllowanceApp.Api/DTO/AllowancePointDTO.cs
@@ -2,7 +2,7 @@ using AllowanceApp.Core.Models;
 
 namespace AllowanceApp.Api.DTO
 {
-    public record AllowancePointDTO(int AccountID, string Category, double Price, double Points, double Total)
+    public record AllowancePointDTO(int AccountID, string Category, int Price, double Points, double Total)
     {
         public AllowancePointDTO(AllowancePoint APoint) : this
         (
