+++ Source/AllowanceApp.Core/Models/AllowancePoint.cs
@@ -1,11 +1,11 @@
 namespace AllowanceApp.Core.Models
 {
-    public class AllowancePoint(string Category, double Price)
+    public class AllowancePoint(string Category, int Price)
     {
         public int AccountID { get; set; } = 0;
         public string Category { get; set; } = Category;
-        public double Price { get; set; } = Price;
+        public int Price { get; set; } = Price;
         public int Points { get; set; } = 0;
-        public double Total => Price * Points;
+        public int Total => Price * Points;
     }
 }
